v2-backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
openpyxl==3.1.5
python-multipart==0.0.20
google-generativeai>=0.8.0,<1.0.0
google-genai>=1.0.0,<2.0.0
google-cloud-documentai>=2.20.0
pdf2image>=1.17.0
Pillow>=10.0.0
openai>=1.0.0,<2.0.0
httpx>=0.24.0
line-bot-sdk>=3.0.0,<4.0.0
slowapi>=0.1.9,<1.0.0
pdfplumber>=0.10.0
ddgs>=9.0.0
